feat(suppliers): add createdAt field and shared date utility
- Add createdAt field to Supplier type and API responses
- Extract formatInsertDate utility to utils/date.ts for reuse in ClientsView
- Add Insert Date column to SuppliersView table following ClientsView pattern
- Update i18n keys for insert date column header
