+++ types.ts
@@ -497,6 +497,7 @@ export interface Supplier {
   taxCode?: string;
   paymentTerms?: string;
   notes?: string;
+  createdAt?: number;
 }
 
 export interface SupplierQuoteItem {
